bugfix: Add GlucoseUpdateService and start at wear main activity (#30)
* bugfix: Add GlucoseUpdateService and start at wear main activity

* Adjust file structure

* Remove unnecessary imports

* Fix build error

* Move shared preference logic into MainActivity and fix type of NightScoutData

* Change update interval to 30 sec

* Use GlucoseUpdateService in KDualCanvasRenderer instead of updating manually

* Update BG when preference change also in watchface

* Sort glucose data for graph

* Remove unused val dummyGraphDexcomData

+++ wear/src/main/java/com/kaist/k_dual/model/Dexcom.kt
@@ -162,7 +162,7 @@ class DexcomClient(configurationProps: ConfigurationProps) {
                     val trend = Trend.valueOf(entry.Trend.uppercase(Locale.ROOT))
                     GlucoseEntry(
                         mgdl = entry.Value,
-                        mmol = mgdlToMmol(entry.Value),
+                        mmol = mgdlToMmol(entry.Value.toDouble()),
                         trend = trend,
                         timestamp = extractNumber(entry.WT)
                     )
@@ -182,7 +182,7 @@ class DexcomClient(configurationProps: ConfigurationProps) {
 
 }
 
-fun mgdlToMmol(mgdl: Int): Double {
+fun mgdlToMmol(mgdl: Double): Double {
     return (mgdl / 18.0).toBigDecimal().setScale(2, RoundingMode.HALF_EVEN).toDouble()
 }
